locations', rooms' and buildings' show pages show title (and description) in the chosen language added some translations
