+++ app/views/rooms/show.html.erb
@@ -1,28 +1,24 @@
 <div class="row">
     <div class="col-md-4">
-        <h2 class="mb-3 fw-light"><%= @room.name %></h2>
+        <h2 class="mb-3 fw-light"> 
+        <%= if I18n.locale == :de
+              @room.name_de
+            elsif I18n.locale == :en
+              @room.name
+            end%></h2>
     </div>
     <div class="col-md-8 align-self-center">
         <% if @room.free? %>
-          <p class="occupancy free">free</p>
+          <p class="occupancy free"> <%=t("free")%></p>
         <% else %>
-          <p class="occupancy occupied">occupied</p>
+          <p class="occupancy occupied"> <%=t("occupied")%></p>
         <% end %>
     </div>
 </div>
 
 <div class="field mb-3">
   <label class="form-label">
-    <strong>Name:</strong>
-  </label>
-  <p class="form-control-static">
-    <%= @room.name %>
-  </p>
-</div>
-
-<div class="field mb-3">
-  <label class="form-label">
-    <strong>Floor:</strong>
+    <strong> <%=t("floor")%>:</strong>
   </label>
   <p class="form-control-static">
     <%= @room.floor %>
@@ -71,10 +67,10 @@
   </ul>
 </div>
 
-<div class="td-btn"><%= link_to 'Calendar', room_calendar_path(@room, start_date: Time.zone.today.beginning_of_week.to_s), class: "btn btn-secondary link-btn" %></div>
+<div class="td-btn"><%= link_to t("calender"), room_calendar_path(@room, start_date: Time.zone.today.beginning_of_week.to_s), class: "btn btn-secondary link-btn" %></div>
 
 <a class="btn btn-primary" href="/route?dest=<%= @room.name %>">
-  <i class="fas fa-route"></i> Directions
+  <i class="fas fa-route"></i> <%= t("directions")%>
 </a>
 
 <script id="_latlonginfo" type="application/json">
