+++ config/locales/de.yml
@@ -15,3 +15,15 @@ de:
     username: 'Benutzername'
     admin?: 'Administrator?'
     user-roles: 'Nutzerrollen'
+    name: 'Name'
+    details: 'Beschreibung'
+    directions: 'Route planen'
+    edit_building: 'Gebäude bearbeiten'
+    free: 'frei'
+    occupied: 'belegt'
+    floor: 'Etage'
+    calender: 'Kalender'
+    openingtimes: 'Öffnungszeiten'
+    back: 'Zurück'
+    show: 'Übersicht'
+
