+++ config/locales/en.yml
@@ -46,3 +46,14 @@ en:
     username: 'User name'
     admin?: 'Admin?'
     user-roles: 'User roles'
+    name: 'Name'
+    details: 'Details'
+    directions: 'Directions'
+    edit_building: 'Edit building'
+    free: 'free'
+    occupied: 'occupied'
+    floor: 'Floor'
+    calender: 'Calender'
+    openingtimes: 'Opening times'
+    back: 'Back'
+    show: 'Show'
